% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contamination_Correction.R
\name{Cal_AUCs}
\alias{Cal_AUCs}
\title{AUROC calculation between the cluster with the lowest expression and each other cluster}
\usage{
Cal_AUCs(object, gene, qualified_cls)
}
\arguments{
\item{object}{a clustered SeuratObject}

\item{gene}{a gene within the input SeuratObject}

\item{qualified_cls}{a vector of qualified cluster names}
}
\value{
a named vector of the AUROC values
}
\description{
Calculate the AUROC values based on a SeuratObject and a selected gene. The cell cluster 
with the lowest expression level will be compared with each other cell cluster (including 
itself) and then generate the corresponding AUROC value. These values will output with 
the cluster names.
}
