% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contamination_Correction.R
\name{Cal_thres}
\alias{Cal_thres}
\title{Find the threshold (best cutting-point) using the Youden index}
\usage{
Cal_thres(object, gene, eGCG_aucs, auc_thres)
}
\arguments{
\item{object}{a clustered SeuratObject}

\item{gene}{a gene within the input SeuratObject}

\item{eGCG_aucs}{a named vector of the AUROC values between the each cluster and the cluster with the lowest expression level}

\item{auc_thres}{the AUROC threshold to determine the boundary between eGCG_positive and eGCG_negative clusters}
}
\value{
the count value of the threshold (best cutting-point)
}
\description{
Calculate the threshold (best cutting-point) based on the Youden indexes between the
count values of the eGCG_postive cluster with the lowest expression and that of all 
eGCG_negative cluster. The best cutting-point will be selected when Younden index is
maximized. The eGCG_positive and eGCG_negative clusters will be identified based on
the input eGCG_aucs and the user-specified auc_thres. The cluster with the auc value
smaller than the auc_thres will be identified as eGCG_negative cluster.
}
\examples{
count_thres <- Cal_thres(seuratobj, gene, eGCG_aucs, auc_thres)
}
