% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationDetection.R
\name{CalAverageExpression.Seurat}
\alias{CalAverageExpression.Seurat}
\title{Calculate average expression}
\usage{
CalAverageExpression.Seurat(object, gene_set, type = "counts")
}
\arguments{
\item{object}{a Seurat object that has been clustered.}

\item{gene_set}{a vector containing genes which you want to calculate average expression.}

\item{type}{a parameter setting for whether to use 'data' slot or 'counts' slot in the Seurat object to calculate average expression
"default" for calculation using 'counts' slot.}
}
\value{
a matrix containing average expression results with gene names as row names,
cluster names as column names.
}
\description{
CalAverageExpression.Seurat calculates average expression of genes within each cluster 
based on the data/counts slot of the given seurat object.
You can learn more about this function at:
}
\examples{
ave.data <- CalAverageExpression.Seurat(object = seuratobj, gene_set = rownames(seuratobj))
}
