% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contamination_Detection.R
\name{CalEnt.Seurat}
\alias{CalEnt.Seurat}
\title{Calculate the entropy}
\usage{
CalEnt.Seurat(object, gene_set)
}
\arguments{
\item{object}{a Seurat object that has been clustered.}

\item{gene_set}{a vector containing genes which you want to calculate entropy for.}
}
\value{
a matrix containing entropy results with gene names as row names, cluster names as column names.
}
\description{
This method calculates the Shannon's entropy of the observed counts of genes within each cluster in the Seurat object.
CalEnt.Seurat calculates entropy of genes based on the 'counts' slot in the Seurat object.
}
\examples{
entropy.data <- CalEnt.Seurat(object = seuratobj, gene_set = rownames(seuratobj))
}
