% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{CalOtsu}
\alias{CalOtsu}
\title{Calculate threshold using Otsu's method}
\usage{
CalOtsu(vec)
}
\arguments{
\item{vec}{a numeric vector containing the whole distribution of counts}
}
\value{
a number indicating threshold calculated by Otsu's method
}
\description{
CalOtsu calculates proper threshold to seperate the distribution into 2 classes
using Otsu's method.
You can learn more about this function at:
}
\examples{
counts <- c(1, 1, 1, 2, 2, 2, 3, 3, 3);thres <- CalOtsu(counts)
}
