% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{ContaminationCorrection}
\alias{ContaminationCorrection}
\title{Perform the contamination correction}
\usage{
ContaminationCorrection(object, cont_genes, auc_thres = 0.9)
}
\arguments{
\item{object}{a clustered SeuratObject}

\item{cont_genes}{a contaminative geneset within the input SeuratObject}

\item{auc_thres}{the AUROC threshold to determine the boundary between eGCG_positive
and eGCG_negative clusters (Default as 0.9, 90%)}
}
\value{
the input clustered SeuratObject with a additional corrected assay of counts
}
\description{
Decontaminate the count matrix of the input SeuratObject based on the input contaminative
genes using a Youden index-based method.
}
\examples{
seuratobj_corrected <- ContaminationCorrection(seuratobj, cont_genes, 
auc_thres = 0.9)
}
