% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contamination_Quantification.R
\name{ContaminationQuantification}
\alias{ContaminationQuantification}
\title{Calculate contamination index for a dataset}
\usage{
ContaminationQuantification(
  object,
  cont_genes,
  auc_thres = 0.9,
  min.cell = 50,
  top10_gcg = F,
  slot = "counts"
)
}
\arguments{
\item{object}{a clustered SeuratObject}

\item{cont_genes}{a contaminative geneset within the input SeuratObject}

\item{min.cell}{the parameter used to filter the cell populations without sufficient number of cells. Cell populations that reaches the threshold could be used in downstream analysis.}

\item{top10_gcg}{a parameter controlling whether to use the top 10 GCGs for calculating contamination index}

\item{slot}{the slot used for calculating contamination index}
}
\value{
the maximum value of the calculated contamination index
}
\description{
Calculating the contamination index for a dataset leveraging GCGs. For each GCG, the contamination level is calculated by dividing its total expression in eGCG- cells by 
the total expression of all genes in the eGCG- cells. The maximum of contamination index is returned. Higher contamination index represents higher contamination level.
}
