% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{correspond_correction}
\alias{correspond_correction}
\title{Correction of contaminated counts based on CDF}
\usage{
correspond_correction(target_cluster_exp, cont_cluster_exp)
}
\arguments{
\item{target_cluster_exp}{numeric vector, the target counts distribution you want to decontaminate}

\item{cont_cluster_exp}{numeric vector, the corresponding contaminative counts distribution}
}
\value{
corrected vector of target count distribution
}
\description{
correspond_correction corrects the contaminated counts based on corresponding
contaminative counts distribution using a CDF-based method
You can learn more about this function at:
}
\examples{
target <- c(5, 6, 7, 8);cont <- c(1, 2, 3, 4);corrected <- correspond_correction(target, cont)
}
