% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{DecontFunc}
\alias{DecontFunc}
\title{Corrects the contamination in Seurat}
\usage{
DecontFunc(object, cont_genes, was.thres)
}
\arguments{
\item{object}{a Seurat object that has been clustered}

\item{cont_genes}{a character vector, containing the genes to be corrected, 
usually generated by ContaminationDetection}

\item{was.thres}{a number indicating the threshold of wasserstein distance for
estimation of containative counts distribution}
}
\value{
a decontaminated count matrix
}
\description{
DecontFunc estimates the contamination of the given genes and corrects them using
Otsu's method or CDF-based method
You can learn more about this function at:
}
\examples{
cont_genes <- c('Ins1', 'Ins2');DecontFunc(object, cont_genes, 1)
}
