% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{Detecting_contaminated_cells}
\alias{Detecting_contaminated_cells}
\title{Estimates the contaminative distribution}
\usage{
Detecting_contaminated_cells(x, cluster_info, was.thres)
}
\arguments{
\item{x}{a named vector, values are normalized counts of a gene, names are the cell barcodes}

\item{cluster_info}{a named vector, names are the cell barcodes, values are the corresponding cluster of the cells,
usually are the active.ident in a Seurat object}

\item{was.thres}{a number indicating the threshold of wasserstein distance for
estimation of contaminative counts distribution}
}
\value{
a character vector containing the barcodes of cells that constitute 
the contamination distribution
}
\description{
Detecting_contaminated_cells find and expand the contaminated cluster to estimate the whole contamination distribution
You can learn more about this function at:
}
\examples{
cont_cells <- Detecting_contaminated_cells(x, Idents(object), 1)
}
