% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationDetection.R
\name{generate_plot}
\alias{generate_plot}
\title{Plot the relationship between the entropy and the expression}
\usage{
generate_plot(.x, name, genes = NULL, point_size = 1.8, cutoff = 0.05)
}
\arguments{
\item{.x}{A dataframe which is the output of the generate_curve function.}

\item{name}{The name printed on the figure.}

\item{genes}{If specified a vector of genes, only these genes would be labeled on the plot. If not specified,
the significantly deviated genes in the dataset would be labeled.}

\item{point_size}{The parameter controls the point size in the plot.}

\item{cutoff}{The threshold controls the justification of significantly deviated genes.}
}
\value{
A ggplot object
}
\description{
generate_plot plots the fitted curve with points. Each point represents a gene in a cluster, the x axis represents the mean expression level
and the y axis represents the entropy.
You can learn more about this function at:
}
\examples{
plot <- generate_plot(result, 'cluster A')
}
