% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{simple_roc}
\alias{simple_roc}
\title{A simple ROC function to obtain the AUROC value}
\usage{
simple_roc(exp, class)
}
\arguments{
\item{exp}{the values representing cellular expression levels of a certain gene}

\item{class}{the cell types of the corresponding cells}
}
\value{
an AUROC (Area Under the Receiver Operating Characteristics) value
}
\description{
Calculate the AUROC based on the values representing cellular expression levels and the
corresponding cell type (two types) information.
}
\examples{
AUC_res <- test_roc(exp_value, cell_type)
}
