% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaminationCorrection.R
\name{ContaminationCorrection}
\alias{ContaminationCorrection}
\title{Interface to use}
\usage{
ContaminationCorrection(object, cont_genes, was.thres = 1)
}
\arguments{
\item{object}{a Seurat object that has been clustered}

\item{cont_genes}{a character vector, containing the genes to be corrected, 
usually generated by ContaminationDetection}

\item{was.thres}{a number indicating the threshold of wasserstein distance for
estimation of containative counts distribution, default is 1}
}
\value{
a decontaminated count matrix
}
\description{
ContaminationCorrection is the interface function to correct the contamination
in single-cell sequencing data
You can learn more about this function at:
}
\examples{
cont_genes <- rownames(ContaminationDetection(object));DecontFunc(object, cont_genes, 1)
}
