% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contamination_Detection.R
\name{generate_plot}
\alias{generate_plot}
\title{Plot the relationship between the entropy and the expression}
\usage{
generate_plot(.x, name, genes = NULL, point_size = 1.8, cutoff = 0.05)
}
\arguments{
\item{.x}{a dataframe which is the output of the generate_curve function.}

\item{name}{the name printed on the figure.}

\item{genes}{if specified a vector of genes, only these genes would be labeled on the plot. If not specified, the significantly deviated genes in the dataset would be labeled.}

\item{point_size}{the parameter controls the point size in the plot.}

\item{cutoff}{the threshold controls the justification of significantly deviated genes.}
}
\value{
A ggplot object
}
\description{
generate_plot plots the fitted curve with points. Each point represents a gene in a cluster, the x axis represents the mean expression level
and the y axis represents the entropy.
}
